requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
python-uinput>=0.11.2
pyautogui>=0.9.54
PyYAML>=6.0
pytest>=7.4.0